MI32: improve dashboard and widget handling
- Preserve fractional values in Berry MI32.set_temp() and MI32.set_hum().
- Limit saved MI32 configurations safely to 32 sensors.
- Respect the configured temperature unit (SetOption8) in dashboard temperature and dew-point values.
- Update browser-side age counters using real elapsed time.
- Allow a Berry widget response to contain multiple independent cards.
- Send dashboard metrics in a dedicated response every ten seconds.
- Pause dashboard polling while the browser tab is hidden.
- Display chained knob button presses as their numeric count.
- Various small improvements.

+++ tasmota/tasmota_xdrv_driver/xdrv_52_3_berry_MI32.ino
@@ -71,11 +71,11 @@ extern "C" {
     return buffer;
   }
 
-  void be_MI32_set_hum(int slot, int hum_val){
+  void be_MI32_set_hum(int slot, float hum_val){
     MI32setHumidityForSlot(slot,hum_val);
   }
 
-  void be_MI32_set_temp(int slot, int temp_val){
+  void be_MI32_set_temp(int slot, float temp_val){
     MI32setTemperatureForSlot(slot,temp_val);
   }
